# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "contact": {
            "email": "nivsoficial@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://localhost:8000/api/tarefas",
            "description": "Tarefas"
        }
    ],
    "paths": {
...
        "/tarefas": {
            "post": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Cria uma nova tarefa",
                "operationId": "426e93026eea745cb2168c27377a5b7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título da tarefa",
                                        "type": "string",
                                        "example": "Tarefa 1"
                                    },
                                    "description": {
                                        "description": "Descrição da tarefa",
                                        "type": "string",
                                        "example": "Descrição da tarefa"
                                    },
                                    "due_date": {
                                        "description": "Data de vencimento da tarefa",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-08-10 15:00:00"
                                    },
                                    "assignee_id": {
                                        "description": "ID do funcionário atribuído à tarefa",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa criada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa criada com sucesso."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Não foi possível criar a tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível criar a tarefa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tarefas/{tarefa_id}": {
            "get": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Retorna os detalhes de uma tarefa",
                "operationId": "53729c3b0b27909708634989292e2d04",
                "parameters": [
                    {
                        "name": "tarefa_id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da tarefa retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sucesso ao encontrar a tarefa."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Não existe essa tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não existe essa tarefa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tarefas/{id}": {
            "put": {
                "tags": [
                    "Tarefas"
                ],
                "summary": "Atualiza os detalhes de uma tarefa",
                "operationId": "70170496b19147355ad48a6ae82a0730",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título da tarefa",
                                        "type": "string",
                                        "example": "Tarefa 1"
                                    },
                                    "description": {
                                        "description": "Descrição da tarefa",
                                        "type": "string",
                                        "example": "Descrição da tarefa"
                                    },
                                    "due_date": {
                                        "description": "Data de vencimento da tarefa",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-08-10 15:00:00"
                                    },
                                    "assignee_id": {
                                        "description": "ID do funcionário atribuído à tarefa",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa atualizada com sucesso."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Não existe essa tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não existe essa tarefa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deleta uma tarefa",
                "operationId": "e0fdf91bbed535a3fbd5bfa06aed9ba5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa deletada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa apagada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Não existe essa tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não existe essa tarefa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}